Update to Forge 13.20.1.2393. This adds rotation origin variable for animated models (#3875). Add ForgeSpinBlock and associated classes based on the updated Forge ModelAnimationDebug.java

+++ src/main/java/tld/testmod/common/animation/ModAnimation.java
@@ -12,7 +12,7 @@ public enum ModAnimation
     /**
      * Get the global world time for the current tick, in seconds.
      */
-    public static float getWorldTime(World world)
+    public static double getWorldTime(World world)
     {
         return getWorldTime(world, 0);
     }
@@ -20,10 +20,10 @@ public enum ModAnimation
     /**
      * Get the global world time for the current tick + partial tick progress, in seconds.
      */
-    public static float getWorldTime(World world, float tickProgress)
+    public static double getWorldTime(World world, float tickProgress)
     {
-        short time = (short)world.getTotalWorldTime();
-        return (time + tickProgress) / 20;
+        long time = (long)world.getTotalWorldTime();
+        return ((double)time + tickProgress) / 20;
     }
 
     /**
